Use logged in user's profile image if applicable
Currently the POST & PUT APIs aren't configured to return a user's
profile image in their response. Since, the currently logged-in user is
the only one that can POST using the app, so, we use the locally stored
profile image in preferences instead.
Incase of PUT we just use the previous profile image version that we had.